# Internal CosmosElements: Adds Perf Benchmarks for Lazy CosmosElements