--- a/Microsoft.Azure.Cosmos/tests/Microsoft.Azure.Cosmos.Performance.Tests/CosmosElements/LazyCosmosElementsBenchmark.cs
+++ b/Microsoft.Azure.Cosmos/tests/Microsoft.Azure.Cosmos.Performance.Tests/CosmosElements/LazyCosmosElementsBenchmark.cs
@@ -1,0 +1,179 @@
+﻿namespace Microsoft.Azure.Cosmos.Performance.Tests.CosmosElements
+{
+    using System;
+    using System.Collections.Generic;
+    using System.Reflection;
+    using BenchmarkDotNet.Attributes;
+    using Microsoft.Azure.Cosmos.CosmosElements;
+    using Microsoft.Azure.Cosmos.Json;
+
+    [MemoryDiagnoser]
+    public class LazyCosmosElementsBenchmarks
+    {
+        public enum AccessPattern
+        {
+            Never,
+            Once,
+            Many
+        }
+
+        public enum SerializationFormat
+        {
+            Text,
+            Binary
+        }
+
+        private static class Payloads
+        {
+            //public static readonly Payload MediumString = Payload.Create(name: "medium string", "\"The quick brown fox jumped over the lazy dog.\"");
+            public static readonly Payload LongString = Payload.Create(name: "long string", "\"Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.\"");
+            //public static readonly Payload Integer = Payload.Create(name: "integer", "1234");
+            public static readonly Payload Double = Payload.Create(name: "double", "1234.5678");
+            //public static readonly Payload Null = Payload.Create(name: "null", "null");
+            //public static readonly Payload True = Payload.Create(name: "true", "true");
+            //public static readonly Payload False = Payload.Create(name: "false", "false");
+            public static readonly Payload Array = Payload.Create(name: "array", "[1, 2, 3, 4, 5, 6, 7, 8, 9, 10]");
+            public static readonly Payload Object = Payload.Create(name: "object", @"{
+                ""id"": ""7029d079-4016-4436-b7da-36c0bae54ff6"",
+                ""double"": 0.18963001816981939,
+                ""int"": -1330192615,
+                ""string"": ""XCPCFXPHHF"",
+                ""boolean"": true,
+                ""null"": null,
+                ""datetime"": ""2526-07-11T18:18:16.4520716"",
+                ""spatialPoint"": {
+                    ""type"": ""Point"",
+                    ""coordinates"": [
+                        118.9897,
+                        -46.6781
+                    ]
+                },
+                ""text"": ""tiger diamond newbrunswick snowleopard chocolate dog snowleopard turtle cat sapphire peach sapphire vancouver white chocolate horse diamond lion superlongcolourname ruby""
+            }");
+        }
+
+        [Benchmark]
+        [ArgumentsSource(nameof(CreateAndNavigateCosmosElementArguments))]
+        public void CreateAndNavigateCosmosElement(
+            Payload payload,
+            SerializationFormat jsonSerializationFormat,
+            AccessPattern accessMode)
+        {
+            ReadOnlyMemory<byte> bytes = jsonSerializationFormat switch
+            {
+                SerializationFormat.Text => payload.Text,
+                SerializationFormat.Binary => payload.Binary,
+                _ => throw new ArgumentOutOfRangeException(nameof(jsonSerializationFormat)),
+            };
+
+            for (int i = 0; i < 1; i++)
+            {
+                CosmosElement element = CosmosElement.CreateFromBuffer(bytes);
+                int numAccessIterations = accessMode switch
+                {
+                    AccessPattern.Never => 0,
+                    AccessPattern.Once => 1,
+                    AccessPattern.Many => 10,
+                    _ => throw new ArgumentOutOfRangeException(nameof(accessMode)),
+                };
+
+                for (int accessIteration = 0; accessIteration < numAccessIterations; accessIteration++)
+                {
+                    Access(element);
+                }
+            }
+        }
+
+        private static void Access(CosmosElement element)
+        {
+            switch (element)
+            {
+                case CosmosBoolean _:
+                case CosmosNull _:
+                    break;
+
+                case CosmosString cosmosString:
+                    _ = cosmosString.Value;
+                    break;
+
+                case CosmosNumber cosmosNumber:
+                    _ = cosmosNumber.Value;
+                    break;
+
+                case CosmosArray cosmosArray:
+                    foreach (CosmosElement arrayItem in cosmosArray)
+                    {
+                        Access(arrayItem);
+                    }
+                    break;
+
+                case CosmosObject cosmosObject:
+                    foreach (KeyValuePair<string, CosmosElement> kvp in cosmosObject)
+                    {
+                        Access(kvp.Value);
+                    }
+                    break;
+
+                default:
+                    throw new ArgumentOutOfRangeException(nameof(element));
+            }
+        }
+
+        public IEnumerable<object[]> CreateAndNavigateCosmosElementArguments()
+        {
+            foreach (FieldInfo fieldInfo in typeof(Payloads).GetFields(BindingFlags.Static | BindingFlags.Public | BindingFlags.NonPublic))
+            {
+                Payload payload = (Payload)fieldInfo.GetValue(null);
+                foreach (SerializationFormat sourceFormat in Enum.GetValues(typeof(SerializationFormat)))
+                {
+                    foreach (AccessPattern accessPattern in Enum.GetValues(typeof(AccessPattern)))
+                    {
+                        yield return new object[] { payload, sourceFormat, accessPattern };
+                    }
+                }
+            }
+        }
+
+        public readonly struct Payload
+        {
+            private Payload(
+                string name,
+                ReadOnlyMemory<byte> text,
+                ReadOnlyMemory<byte> binary)
+            {
+                this.Name = name;
+                this.Text = text;
+                this.Binary = binary;
+            }
+
+            public string Name { get; }
+            public ReadOnlyMemory<byte> Text { get; }
+            public ReadOnlyMemory<byte> Binary { get; }
+
+            public static Payload Create(string name, string jsonText)
+            {
+                if (jsonText == null)
+                {
+                    throw new ArgumentNullException(nameof(jsonText));
+                }
+
+                CosmosElement element = CosmosElement.Parse(jsonText);
+                IJsonWriter jsonBinaryWriter = JsonWriter.Create(JsonSerializationFormat.Binary);
+                IJsonWriter jsonTextWriter = JsonWriter.Create(JsonSerializationFormat.Text);
+
+                element.WriteTo(jsonTextWriter);
+                element.WriteTo(jsonBinaryWriter);
+
+                ReadOnlyMemory<byte> text = jsonTextWriter.GetResult();
+                ReadOnlyMemory<byte> binary = jsonBinaryWriter.GetResult();
+
+                return new Payload(name, text, binary);
+            }
+
+            public override string ToString()
+            {
+                return this.Name;
+            }
+        }
+    }
+}