# Internal CosmosElements: Adds Perf Benchmarks for Lazy CosmosElements

## Description

Adds perf benchmarks for LazyCosmosElements. The general benchmark will create a cosmos element of a particular json type using a particular serialization format and access it with a particular pattern. For now we have the cross product of the following:

### JsonTypes
* Array
* Object
* Double
* String

### SerializationFormat
* Text
* Binary

### Access Pattern
* None
* Once
* Many

and got the following results:

```
|                         Method |       payload | jsonSerializationFormat | accessMode |        Mean |       Error |      StdDev |   Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------------------------- |-------------- |------------------------ |----------- |------------:|------------:|------------:|--------:|------:|------:|----------:|
| CreateAndNavigateCosmosElement |         array |                    Text |      Never |    181.9 ns |     4.14 ns |     4.06 ns |  0.0975 |     - |     - |     408 B |
| CreateAndNavigateCosmosElement |         array |                    Text |       Once |  4,985.5 ns |    58.08 ns |    54.33 ns |  1.1902 |     - |     - |    5000 B |
| CreateAndNavigateCosmosElement |         array |                    Text |       Many |  7,435.5 ns |   127.38 ns |   106.37 ns |  1.2970 |     - |     - |    5432 B |
| CreateAndNavigateCosmosElement |         array |                  Binary |      Never |    158.4 ns |     3.00 ns |     3.21 ns |  0.0629 |     - |     - |     264 B |
| CreateAndNavigateCosmosElement |         array |                  Binary |       Once |  3,352.6 ns |    42.30 ns |    37.50 ns |  1.1826 |     - |     - |    4960 B |
| CreateAndNavigateCosmosElement |         array |                  Binary |       Many |  5,764.0 ns |    96.61 ns |    90.37 ns |  1.2817 |     - |     - |    5392 B |
| CreateAndNavigateCosmosElement |        double |                    Text |      Never |    568.5 ns |     7.28 ns |     6.45 ns |  0.1183 |     - |     - |     496 B |
| CreateAndNavigateCosmosElement |        double |                    Text |       Once |    714.0 ns |     3.63 ns |     3.22 ns |  0.1183 |     - |     - |     496 B |
| CreateAndNavigateCosmosElement |        double |                    Text |       Many |    813.2 ns |     9.39 ns |     8.79 ns |  0.1183 |     - |     - |     496 B |
| CreateAndNavigateCosmosElement |        double |                  Binary |      Never |    161.8 ns |     2.23 ns |     1.98 ns |  0.0725 |     - |     - |     304 B |
| CreateAndNavigateCosmosElement |        double |                  Binary |       Once |    242.9 ns |     2.38 ns |     2.23 ns |  0.0725 |     - |     - |     304 B |
| CreateAndNavigateCosmosElement |        double |                  Binary |       Many |    424.4 ns |    14.01 ns |    40.65 ns |  0.0725 |     - |     - |     304 B |
| CreateAndNavigateCosmosElement |   long string |                    Text |      Never |  1,668.1 ns |    33.34 ns |    34.24 ns |  0.1087 |     - |     - |     456 B |
| CreateAndNavigateCosmosElement |   long string |                    Text |       Once |  2,077.6 ns |    41.35 ns |    42.47 ns |  0.3242 |     - |     - |    1368 B |
| CreateAndNavigateCosmosElement |   long string |                    Text |       Many |  2,166.3 ns |    42.84 ns |    55.71 ns |  0.3242 |     - |     - |    1368 B |
| CreateAndNavigateCosmosElement |   long string |                  Binary |      Never |    161.4 ns |     2.78 ns |     2.60 ns |  0.0629 |     - |     - |     264 B |
| CreateAndNavigateCosmosElement |   long string |                  Binary |       Once |  1,383.4 ns |    20.87 ns |    19.52 ns |  0.2804 |     - |     - |    1176 B |
| CreateAndNavigateCosmosElement |   long string |                  Binary |       Many |  1,445.7 ns |    14.00 ns |    12.41 ns |  0.2804 |     - |     - |    1176 B |
| CreateAndNavigateCosmosElement |        object |                    Text |      Never |    331.8 ns |     6.65 ns |    11.99 ns |  0.2637 |     - |     - |    1104 B |
| CreateAndNavigateCosmosElement |        object |                    Text |       Once | 12,262.7 ns |    73.82 ns |    57.63 ns |  1.7700 |     - |     - |    7408 B |
| CreateAndNavigateCosmosElement |        object |                    Text |       Many | 53,472.7 ns | 1,049.10 ns | 1,208.14 ns | 11.5967 |     - |     - |   48736 B |
| CreateAndNavigateCosmosElement |        object |                  Binary |      Never |    293.9 ns |     5.92 ns |     8.85 ns |  0.2294 |     - |     - |     960 B |
| CreateAndNavigateCosmosElement |        object |                  Binary |       Once |  8,145.3 ns |   153.00 ns |   135.63 ns |  1.8158 |     - |     - |    7608 B |
| CreateAndNavigateCosmosElement |        object |                  Binary |       Many | 76,662.6 ns |   950.29 ns |   888.90 ns | 16.1133 |     - |     - |   67440 B |
```

From here we can see that we are doing some things well and some things horribly. 

In terms of good things we can see that the lazy aspect of CosmosElement is being honored, since an object takes about the same amount of time to create as a number. We can also see that repeated accesses to numbers and strings does not cost us additional time or memory.

In terms of bad things ... it appears that CosmosObject and CosmosArray are not cacheing results well with CosmosObject looking like it does not at all. These will have to be investigated.

Future improvements:
* Combine navigator and node into a single type to save on the number of references
* No need to store navigator and node since the lambda for the lazy type already captured it.
* Consider adding a lazy to eager API so that people can materialize their CosmosElements and also for GC to reclaim the difference in space between the two APIs.